# Json : Refactor JsonTextNavigator Performance Improvements