--- a/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBReader.cs
+++ b/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBReader.cs
@@ -97,43 +97,43 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
             {
                 switch (this.reader.TokenType)
                 {
-                    case Newtonsoft.Json.JsonToken.None:
-                    case Newtonsoft.Json.JsonToken.StartConstructor:
-                    case Newtonsoft.Json.JsonToken.EndConstructor:
-                        throw new InvalidOperationException();
                     case Newtonsoft.Json.JsonToken.StartObject:
                         this.JsonObjectState.RegisterToken(JsonTokenType.BeginObject);
                         break;
+
                     case Newtonsoft.Json.JsonToken.StartArray:
                         this.JsonObjectState.RegisterToken(JsonTokenType.BeginArray);
                         break;
+
                     case Newtonsoft.Json.JsonToken.PropertyName:
                         this.JsonObjectState.RegisterToken(JsonTokenType.FieldName);
                         break;
-                    case Newtonsoft.Json.JsonToken.Comment:
-                    case Newtonsoft.Json.JsonToken.Raw:
+
                     case Newtonsoft.Json.JsonToken.String:
-                    case Newtonsoft.Json.JsonToken.Date:
-                    case Newtonsoft.Json.JsonToken.Bytes:
                         this.JsonObjectState.RegisterToken(JsonTokenType.String);
                         break;
+
                     case Newtonsoft.Json.JsonToken.Integer:
                     case Newtonsoft.Json.JsonToken.Float:
                         this.JsonObjectState.RegisterToken(JsonTokenType.Number);
                         break;
+
                     case Newtonsoft.Json.JsonToken.Boolean:
-                        this.JsonObjectState.RegisterToken(this.reader.Value.ToString() == true.ToString() ? JsonTokenType.True : JsonTokenType.False);
+                        this.JsonObjectState.RegisterToken(this.reader.Value is true ? JsonTokenType.True : JsonTokenType.False);
                         break;
+
                     case Newtonsoft.Json.JsonToken.Null:
-                    case Newtonsoft.Json.JsonToken.Undefined:
                         this.JsonObjectState.RegisterToken(JsonTokenType.Null);
                         break;
+
                     case Newtonsoft.Json.JsonToken.EndObject:
                         this.JsonObjectState.RegisterToken(JsonTokenType.EndObject);
                         break;
+
                     case Newtonsoft.Json.JsonToken.EndArray:
                         this.JsonObjectState.RegisterToken(JsonTokenType.EndArray);
                         break;
+
                     default:
                         throw new ArgumentException("Got an invalid newtonsoft type");
                 }
@@ -155,8 +155,12 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
             }
 
             StreamReader streamReader = new StreamReader(stream, Encoding.UTF8);
-            Newtonsoft.Json.JsonTextReader newtonsoftReader = new Newtonsoft.Json.JsonTextReader(streamReader);
-            return new NewtonsoftToCosmosDBReader(newtonsoftReader);
+            Newtonsoft.Json.JsonTextReader newtonsoftReader = new Newtonsoft.Json.JsonTextReader(streamReader)
+            {
+                DateParseHandling = Newtonsoft.Json.DateParseHandling.None
+            };
+
+            return NewtonsoftToCosmosDBReader.CreateFromReader(newtonsoftReader);
         }
 
         public static NewtonsoftToCosmosDBReader CreateFromString(string json)
@@ -167,8 +171,12 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
             }
 
             StringReader stringReader = new StringReader(json);
-            Newtonsoft.Json.JsonTextReader newtonsoftReader = new Newtonsoft.Json.JsonTextReader(stringReader);
-            return new NewtonsoftToCosmosDBReader(newtonsoftReader);
+            Newtonsoft.Json.JsonTextReader newtonsoftReader = new Newtonsoft.Json.JsonTextReader(stringReader)
+            {
+                DateParseHandling = Newtonsoft.Json.DateParseHandling.None
+            };
+
+            return NewtonsoftToCosmosDBReader.CreateFromReader(newtonsoftReader);
         }
 
         public static NewtonsoftToCosmosDBReader CreateFromReader(Newtonsoft.Json.JsonReader reader)