--- a/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBReader.cs
+++ b/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBReader.cs
@@ -17,7 +17,6 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
         private readonly Newtonsoft.Json.JsonReader reader;
 
         private NewtonsoftToCosmosDBReader(Newtonsoft.Json.JsonReader reader)
-            : base(true)
         {
             this.reader = reader ?? throw new ArgumentNullException(nameof(reader));
         }