--- a/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBWriter.cs
+++ b/Microsoft.Azure.Cosmos/src/Json/Interop/NewtonsoftToCosmosDBWriter.cs
@@ -7,6 +7,7 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
     using System.IO;
     using System.Text;
     using Microsoft.Azure.Cosmos.Core.Utf8;
+    using Newtonsoft.Json;
 
 #if INTERNAL
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
@@ -146,7 +147,11 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
             ReadOnlySpan<byte> rawJsonToken)
         {
             string rawJson = Encoding.UTF8.GetString(rawJsonToken);
-            Newtonsoft.Json.JsonTextReader jsonTextReader = new Newtonsoft.Json.JsonTextReader(new StringReader(rawJson));
+            Newtonsoft.Json.JsonTextReader jsonTextReader = new Newtonsoft.Json.JsonTextReader(new StringReader(rawJson))
+            {
+                DateParseHandling = DateParseHandling.None,
+            };
+
             while (jsonTextReader.Read())
             {
                 if (jsonTokenType == JsonTokenType.FieldName)
@@ -155,8 +160,44 @@ namespace Microsoft.Azure.Cosmos.Json.Interop
                 }
                 else
                 {
-                    this.writer.WriteValue(jsonTextReader.Value);
+                    switch (jsonTextReader.TokenType)
+                    {
+                        case Newtonsoft.Json.JsonToken.StartObject:
+                            this.writer.WriteStartObject();
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.StartArray:
+                            this.writer.WriteStartArray();
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.PropertyName:
+                            this.writer.WritePropertyName(jsonTextReader.Value as string);
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.Integer:
+                        case Newtonsoft.Json.JsonToken.Float:
+                        case Newtonsoft.Json.JsonToken.String:
+                        case Newtonsoft.Json.JsonToken.Boolean:
+                            this.writer.WriteValue(jsonTextReader.Value);
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.Null:
+                            this.writer.WriteNull();
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.EndObject:
+                            this.writer.WriteEndObject();
+                            break;
+
+                        case Newtonsoft.Json.JsonToken.EndArray:
+                            this.writer.WriteEndArray();
+                            break;
+
+                        default:
+                            throw new ArgumentOutOfRangeException($"Unknown {nameof(JsonToken)}: {jsonTextReader.TokenType}.");
+                    }
                 }
+                
             }
         }