--- a/Microsoft.Azure.Cosmos/src/Json/JsonNavigator.JsonBinaryNavigator.cs
+++ b/Microsoft.Azure.Cosmos/src/Json/JsonNavigator.JsonBinaryNavigator.cs
@@ -5,11 +5,8 @@ namespace Microsoft.Azure.Cosmos.Json
 {
     using System;
     using System.Collections.Generic;
-    using System.IO;
     using System.Runtime.InteropServices;
-    using System.Text;
     using Microsoft.Azure.Cosmos.Core.Utf8;
-    using Microsoft.Azure.Cosmos.Linq;
 
     /// <summary>
     /// Partial class that wraps the private JsonTextNavigator
@@ -35,11 +32,9 @@ namespace Microsoft.Azure.Cosmos.Json
             /// </summary>
             /// <param name="buffer">The (UTF-8) buffer to navigate.</param>
             /// <param name="jsonStringDictionary">The JSON string dictionary.</param>
-            /// <param name="skipValidation">whether to skip validation or not.</param>
             public JsonBinaryNavigator(
                 ReadOnlyMemory<byte> buffer,
-                JsonStringDictionary jsonStringDictionary,
-                bool skipValidation = false)
+                JsonStringDictionary jsonStringDictionary)
             {
                 if (buffer.Length < 2)
                 {