--- a/Microsoft.Azure.Cosmos/src/Json/JsonNavigator.cs
+++ b/Microsoft.Azure.Cosmos/src/Json/JsonNavigator.cs
@@ -33,12 +33,10 @@ namespace Microsoft.Azure.Cosmos.Json
         /// </summary>
         /// <param name="buffer">The buffer to navigate</param>
         /// <param name="jsonStringDictionary">The optional json string dictionary for binary encoding.</param>
-        /// <param name="skipValidation">Whether validation should be skipped.</param>
         /// <returns>A concrete JsonNavigator that can navigate the supplied buffer.</returns>
         public static IJsonNavigator Create(
             ReadOnlyMemory<byte> buffer,
-            JsonStringDictionary jsonStringDictionary = null,
-            bool skipValidation = false)
+            JsonStringDictionary jsonStringDictionary = null)
         {
             if (buffer.IsEmpty)
             {
@@ -52,10 +50,10 @@ namespace Microsoft.Azure.Cosmos.Json
             {
                 // Explicitly pick from the set of supported formats
                 case JsonSerializationFormat.Binary:
-                    return new JsonBinaryNavigator(buffer, jsonStringDictionary, skipValidation);
+                    return new JsonBinaryNavigator(buffer, jsonStringDictionary);
                 default:
                     // or otherwise assume text format
-                    return new JsonTextNavigator(buffer, skipValidation);
+                    return new JsonTextNavigator(buffer);
             }
         }