… Microsoft.Azure.Cosmos/src/Json/JsonReader.cs
… Microsoft.Azure.Cosmos/src/Json/JsonSerializer.cs
… Microsoft.Azure.Cosmos/src/Json/JsonWriter.JsonBinaryWriter.cs
… Microsoft.Azure.Cosmos/src/Json/JsonWriter.JsonTextWriter.cs
… Microsoft.Azure.Cosmos/src/Json/JsonWriter.cs
… Microsoft.Azure.Cosmos/src/Json/Utf8Memory.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/CosmosItemTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/Query/SanityQueryTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/CosmosElements/DeserializerBenchmark.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/JsonRoundtripBenchmark.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/JsonRoundtripBenchmark.tt
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/JsonRoundtripBenchmarkResults.txt
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/SerializerBenchmark.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/DeserializerTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonNewtonsoftNavigator.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonRoundtripTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonTestUtils.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/NewtonsoftInteropTests.cs
… Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/SerializerTests.cs