# Json : JsonTextNavigator Performance Improvements

## Description

`JsonTextNavigator` was not performing a memcpy when writing to a `JsonTextWriter`. Instead it was traversing a DOM and writing out each token one at a time, which requires materializing said tokens. This hurts the stream story, since it reads the REST response from the backend and writes it out to a stream one document at a time. This required a deserialization + serialization, when only a memcpy was needed.

The optimization was done in two phases: 

1) Implement `bool JsonTextNavigator.TryGetBufferedRawJson(out ReadOnlyMemory<byte>)` which allows the navigator to avoid serializing the results to the the writer one token at a time (and thus materializing individual numbers and strings).

2) Add `LazyNode`, which defers forming the AST, until the user actually wants to navigate (instead of just writing the node out to a buffer). This allows the navigator to avoid the partial deserialization needed to form the AST of the JSON. 

The perf results are as follows:

```
Original:

|                                                     Method |         Mean |      Error |      StdDev |       Median |      Gen 0 |     Gen 1 | Gen 2 |    Allocated |
|----------------------------------------------------------- |-------------:|-----------:|------------:|-------------:|-----------:|----------:|------:|-------------:|
|                  NutritionData_TextNavigator_To_TextWriter | 44,110.66 us | 860.374 us | 1,364.64 us | 43,695.80 us |  2000.0000 | 1000.0000 |     - |  12459.99 KB |
|                NutritionData_TextNavigator_To_BinaryWriter | 38,175.24 us | 748.318 us | 1,229.51 us | 37,846.00 us |  2000.0000 | 1000.0000 |     - |  12575.37 KB |
|                NutritionData_BinaryNavigator_To_TextWriter | 17,181.55 us | 387.847 us |   490.50 us | 17,303.10 us |          - |         - |     - |   3564.78 KB |
|              NutritionData_BinaryNavigator_To_BinaryWriter |     65.22 us |   5.918 us |    17.45 us |     64.50 us |          - |         - |     - |    687.36 KB |

Enabled TextNavigator.TryGetBufferedRawJson:

|                                                     Method |         Mean |        Error |      StdDev |       Median |      Gen 0 |     Gen 1 | Gen 2 |    Allocated |
|----------------------------------------------------------- |-------------:|-------------:|------------:|-------------:|-----------:|----------:|------:|-------------:|
|                  NutritionData_TextNavigator_To_TextWriter | 10,693.98 us |   212.918 us |   467.36 us | 10,612.45 us |          - |         - |     - |   3949.94 KB |
|                NutritionData_TextNavigator_To_BinaryWriter | 20,694.09 us |   407.582 us |   976.54 us | 20,425.80 us |          - |         - |     - |   3959.93 KB |
|                NutritionData_BinaryNavigator_To_TextWriter | 17,248.85 us |   354.927 us |   461.50 us | 17,190.50 us |          - |         - |     - |   3564.78 KB |
|              NutritionData_BinaryNavigator_To_BinaryWriter |     61.17 us |     6.273 us |    18.30 us |     56.50 us |          - |         - |     - |    687.36 KB |

Lazy Root Node: 

|                                                     Method |         Mean |        Error |      StdDev |       Median |      Gen 0 |     Gen 1 | Gen 2 |    Allocated |
|----------------------------------------------------------- |-------------:|-------------:|------------:|-------------:|-----------:|----------:|------:|-------------:|
|                  NutritionData_TextNavigator_To_TextWriter |    126.38 us |    10.785 us |    31.80 us |    123.50 us |          - |         - |     - |   1522.67 KB |
|                NutritionData_TextNavigator_To_BinaryWriter | 20,321.63 us |   401.572 us |   648.46 us | 20,112.65 us |          - |         - |     - |   3960.15 KB |
|                NutritionData_BinaryNavigator_To_TextWriter | 17,373.66 us |   338.177 us |   506.17 us | 17,291.20 us |          - |         - |     - |   3564.78 KB |
|              NutritionData_BinaryNavigator_To_BinaryWriter |     62.16 us |     6.348 us |    18.52 us |     58.50 us |          - |         - |     - |    687.36 KB |
```

Which bring the overall time from `44,110.66 us` down to `126.38 us` -> `350x improvement` or about `1% of the original time`.